+++ network/src/commonMain/kotlin/com/wire/kalium/network/api/v5/authenticated/MLSMessageApiV5.kt
@@ -19,8 +19,40 @@
 package com.wire.kalium.network.api.v5.authenticated
 
 import com.wire.kalium.network.AuthenticatedNetworkClient
+import com.wire.kalium.network.api.base.authenticated.message.MLSMessageApi
+import com.wire.kalium.network.api.base.authenticated.message.SendMLSMessageResponse
 import com.wire.kalium.network.api.v4.authenticated.MLSMessageApiV4
+import com.wire.kalium.network.serialization.Mls
+import com.wire.kalium.network.utils.NetworkResponse
+import com.wire.kalium.network.utils.wrapKaliumResponse
+import io.ktor.client.request.post
+import io.ktor.client.request.setBody
+import io.ktor.http.ContentType
+import io.ktor.http.contentType
 
 internal open class MLSMessageApiV5 internal constructor(
     private val authenticatedNetworkClient: AuthenticatedNetworkClient
-) : MLSMessageApiV4(authenticatedNetworkClient)
+) : MLSMessageApiV4() {
+    private val httpClient get() = authenticatedNetworkClient.httpClient
+
+    override suspend fun sendMessage(message: MLSMessageApi.Message): NetworkResponse<SendMLSMessageResponse> =
+        wrapKaliumResponse {
+            httpClient.post(PATH_MESSAGE) {
+                setBody(message.value)
+                contentType(ContentType.Message.Mls)
+            }
+        }
+
+    override suspend fun sendCommitBundle(bundle: MLSMessageApi.CommitBundle): NetworkResponse<SendMLSMessageResponse> =
+        wrapKaliumResponse {
+            httpClient.post(PATH_COMMIT_BUNDLES) {
+                setBody(bundle.value)
+                contentType(ContentType.Message.Mls)
+            }
+        }
+
+    private companion object {
+        const val PATH_COMMIT_BUNDLES = "mls/commit-bundles"
+        const val PATH_MESSAGE = "mls/messages"
+    }
+}
